+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 
-### 0.0.3
+## 0.0.4
+* Added keyboard shortcut to build and run: `shift+f6`
+* Added keyboard shortcut to build: `ctrl+shift+f6`
+* Added keyboard shortcut to run: `shift+f2`
+* Added `Bar: Edit config` command.
+
+## 0.0.3
 * When the build fails, show the output in Bar Output.
 * Ask before creating bar.conf.json
 * Autoload the extension if `.vscode/bar.conf.json` exists.
@@ -10,9 +16,9 @@
 * Added more info in `package.json`.
 * Added more bugs.
 
-### 0.0.2
+## 0.0.2
 * Updated README.md  
 * Added `Bar: Reset` command.
 
-### 0.0.1
+## 0.0.1
 * Initial release of Bar!